Removed delete button if not admin

+++ recorder2/controller/public/js/recording-list.js
@@ -14,7 +14,7 @@ function RecordingList(root, navItem, hashTracker) {
 
 RecordingList.prototype.onActivate = function() {
     this.hashTracker.setAction("recording-list");
-    
+
     if (!this.isLoaded)
         this.refreshList();
 }
@@ -28,7 +28,7 @@ RecordingList.prototype.onHashChanged = function(action, params) {
 RecordingList.prototype.deleteRecording = function(recordingName) {
     var view = this;
     beginLoadingModal();
-    
+
     $.getJSON(BACKEND_URL + "/recordings/delete?name=" + recordingName,
         function(result) {
             if (result.result) {
@@ -38,8 +38,8 @@ RecordingList.prototype.deleteRecording = function(recordingName) {
             } else {
                 toastr["error"](result.error, "Delete Recording Failed");
             }
-            
-            endLoadingModal();            
+
+            endLoadingModal();
         }
     ).fail(function() {
         toastr["error"]("Failed to delete recording");
@@ -52,7 +52,7 @@ RecordingList.prototype.onOpenRecordingButtonClicked = function(recordingName) {
 }
 
 RecordingList.prototype.onDeleteRecordingButtonClicked = function(recordingName) {
-    var view = this;   
+    var view = this;
     bootbox.dialog({
         message: "Are you sure you want to delete the recording '" + recordingName + "'?<br>This action permanent and cannot be undone.",
         title: "Confirm Deletion",
@@ -78,7 +78,7 @@ RecordingList.prototype.refreshList = function() {
     $.getJSON(BACKEND_URL + "/recordings/list",
         function (data) {
             //console.log("recording list", data);
-            
+
             // Empty current list
             var tbody = view.listTable.find("tbody");
             tbody.empty();
@@ -92,13 +92,13 @@ RecordingList.prototype.refreshList = function() {
             // Output table of records
             $.each(data, function(key, value) {
                 var recordingName = value.name;
-                
+
                 var row = $("<tr>");
                 row.attr("data-recording-name", recordingName);
-                
+
                 // Date/time
                 var startTime = value.time;
-                var endTime = moment(value.time).add(value.length, "seconds").local(); 
+                var endTime = moment(value.time).add(value.length, "seconds").local();
                 row.append($("<td>").text(startTime.format(DATE_DISPLAY_FORMAT) + ", " + startTime.format(TIME_DISPLAY_FORMAT) + " - " + endTime.format(TIME_DISPLAY_FORMAT)));
                 row.append($("<td>").text(value.name));
                 row.append($("<td>").text(value.title));
@@ -109,19 +109,21 @@ RecordingList.prototype.refreshList = function() {
 
                 var buttonColumn = $("<td>");
                 buttonColumn.attr("class", "text-right");
-                
+
                 var viewButton = $("<button>").attr("class", "btn btn-sm btn-primary xs-left-margin xs-bottom-margin").html("<i class='glyphicon glyphicon-folder-open sm-right-margin'></i>View");
                 viewButton.click(function() { view.onOpenRecordingButtonClicked(recordingName); });
                 viewButton.appendTo(buttonColumn);
-                var downloadURL = BACKEND_URL + "/recordings/download?name=" + recordingName; 
+                var downloadURL = BACKEND_URL + "/recordings/download?name=" + recordingName;
                 var downloadButton = $("<a>").attr("class", "btn btn-sm btn-info xs-left-margin xs-bottom-margin").html("<i class='glyphicon glyphicon-download-alt xs-right-margin'></i>Download");
                 downloadButton.attr("href", downloadURL);
                 downloadButton.attr("download", "");
                 downloadButton.appendTo(buttonColumn);
-                var deleteButton = $("<button>").attr("class", "btn btn-sm btn-danger xs-left-margin xs-bottom-margin").html("<i class='glyphicon glyphicon-remove xs-right-margin'></i>Delete");
-                deleteButton.click(function() { view.onDeleteRecordingButtonClicked(recordingName); });
-                deleteButton.appendTo(buttonColumn);
-                
+                if(sessionStorage.getItem("isAdmin") === 1){
+                  var deleteButton = $("<button>").attr("class", "btn btn-sm btn-danger xs-left-margin xs-bottom-margin").html("<i class='glyphicon glyphicon-remove xs-right-margin'></i>Delete");
+                  deleteButton.click(function() { view.onDeleteRecordingButtonClicked(recordingName); });
+                  deleteButton.appendTo(buttonColumn);
+                }
+
                 row.append(buttonColumn);
 
                 tbody.append(row);
@@ -135,4 +137,3 @@ RecordingList.prototype.refreshList = function() {
         endLoadingModal();
     });
 }
-
